check uniqs user and check length

+++ src/main/java/com/example/realinstagram/servises/PostImpl.java
@@ -1,6 +1,7 @@
 package com.example.realinstagram.servises;
 
 import com.example.realinstagram.daos.PostDao;
+import com.example.realinstagram.exceptions.NotFound;
 import com.example.realinstagram.interfaces.PostInterface;
 import com.example.realinstagram.models.Post;
 import com.example.realinstagram.models.User;
@@ -19,23 +20,25 @@ public class PostImpl implements PostInterface {
     @Override
     public String savePosts(User currentUser, String post) {
         try {
-            postDao.addPost(currentUser, post);
-            return "successfully added";
-        } catch (SQLException e) {
-            throw new RuntimeException(e);
+            if (checkPosts(post)) {
+                postDao.addPost(currentUser, post);
+                return "successfully added";
+            } else throw new NotFound("post is empty");
+        } catch (SQLException | NotFound e) {
+            return e.getMessage();
         }
     }
-
     @Override
     public String deletePost(Long id) {
         try {
-            postDao.deletePost(id);
-            return "successfully deleted";
-        } catch (SQLException e) {
-            throw new RuntimeException(e);
+            if (checkPostsById(id)) {
+                postDao.deletePost(id);
+                return "successfully deleted";
+            } else throw new NotFound("post with id: '"+id+"' not found");
+        } catch (SQLException | NotFound e) {
+            return e.getMessage();
         }
     }
-
     @Override
     public List<Post> getAllPostInUser(User user) {
         try {
@@ -44,31 +47,40 @@ public class PostImpl implements PostInterface {
             throw new RuntimeException(e);
         }
     }
-
     @Override
     public List<Post> findPostsByName(String title) {
         try {
-           return postDao.findByName(title);
-        } catch (SQLException e) {
+            if (checkPosts(title)) {
+                return postDao.findByName(title);
+            } else throw new NotFound("post with name: '"+title+"' not found");
+        } catch (SQLException | NotFound e) {
             throw new RuntimeException(e);
         }
     }
-
     @Override
     public Post findPostById(Long id) {
         try {
-            return postDao.findById(id);
+            if (checkPostsById(id)) {
+                return postDao.findById(id);
+            } else throw new NotFound("post with id: '"+id+"' not found");
+        } catch (SQLException | NotFound e) {
+            throw new RuntimeException(e);
+        }
+    }
+    @Override
+    public List<Post> findPostByUser(User user) {
+        try {
+            return postDao.getAllPostInUser(user);
         } catch (SQLException e) {
             throw new RuntimeException(e);
         }
     }
+    private boolean checkPosts(String post) {
+        return !post.isEmpty();
+    }
+    private boolean checkPostsById(Long id) throws SQLException {
+        return postDao.getAllPosts().stream()
+                .anyMatch(post -> post.getId().equals(id));
 
-    @Override
-    public List<Post> findPostByUser(User user) {
-        try {
-           return postDao.getAllPostInUser(user);
-        } catch (SQLException e) {
-            throw new RuntimeException(e);
-        }
     }
 }
